Initial checkin of basic Room datatypes.
I started trying to build something like a linked list of Rooms. This
is naive; since Haskell doesn't have pointers, building the list of
Rooms gets ugly in a hurry and inspecting them is ugly, too.

Instead, we have a lookup table of Rooms and we map a Room to all the
exits from that Room.

Linking two rooms bidirectionally will be really common, so makeLink
is a helper method which generates the correct Map entries.

It may be that this method needs to interact with the world map for
it to make sense, or that we'll need a method to wrap Map.insert, as
neither makeLink nor Map.insert have any awareness of any pre-existing
exits. Otherwise any insert would stomp on any extant exits if
you're, say, trying to update an existing Room's list of exits.